Move test cases up to the adapter level where they now belong.
